requirements.txt
tcolorpy
adafruit_circuitpython_neopixel
adafruit_circuitpython_pixelbuf
pyyaml
fastapi
loguru